Replace the passed random on constructor by new one
This does not change the behavior due to code on KnapsackSimulatedAnnealing on line 43. The .nextDouble() function never runs because in the test case always the new found solutions is better, so being an or this expression never is evaluated.
For this reason changing the generator does not change the behavior nor the test output.